refactor: move gravity mediums to the same folder as my collection mediums
